Conditionally add TLS when connecting to a store
The grpc-client-dns-server-name needs to propogate to the storeset so that
TLS is conditionally added to the grpc dialOpts based on whether it is set.

Test that the grpc-client-tls-secure flag is also set when the
grpc-client-dns-server-name flag is set to make it clear that a secure store
connection is required when using the latter flag.

Signed-off-by: JP Sullivan <jonpsull@cisco.com>