compact: unwrap the original error as well
Just like in IsHaltError(), let's unwrap the given error to see if is
`terrors.MultiError`. This lets us detect the retryable errors as well:

* `BucketCompactor.Compact` returns `terrors.MultiError` of wrapped
errors;
* `compactMainFn` wraps the previously returned error again.

So, we need to unwrap it at the beginning, iterate over it, and then
unwrap again which is what the code does with this change.

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>