compact/tools: add hidden option for retrying downloads
In certain environments, especially private, on-premise clouds,
sometimes transient errors can happen due to misconfiguration or some
reason. In my case, the S3 end-point sometimes had spuriously reset the TCP
connection. Without functionality like this, it was almost impossible
for Thanos Compact to finish its work considering sometimes we have to
download hundreds of gigabytes of files.

Curiously enough, only the downloading path was affected so this only
adds retries for that direction.

Opening this up without any tests to see if it is something we want to
add to Thanos or we should implement
https://github.com/thanos-io/thanos/issues/2784? Or maybe both?

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>