Query: always return a string in the `lastError` field (#2809)
* Query: always return a string in the `lastError` field

While testing out the new React UI, I have found out that some errors
return empty dicts when marshalled into the JSON format. The response
looks like this:

```
{..., "lastError": {}}
```

And then, obviously, the UI fails to parse that. Add a
`stringifiedError` type with tests which ensures that we always get a
string.

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>

* CHANGELOG: add PR number + remove whitespace

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>

* query/storeset: fix a subtle bug

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>

* Make changes according to Bartek's suggestions

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>