sidecar: fix sidecar crashing with lstat errors during a reload from a configmap update
During a reload it's possible that files are added/removed while
we walk the directory and we try to lstat a file that doesn't exist.

This is especially common if running in k8s with the configuration mounted
as a configmap, and the configmap contents change.

With this change we will retry for a short while if we hit any
filesystem errors during the reload while trying to calculate the hash.

Signed-off-by: Ergin Babani <ergin@autonomic.ai>