Disable Azure blob exception logging
Ever since the introduction of the deletion marker, Thanos compact has
been filling the logs with Azure blob storage exceptions. As discussed
in #2565, the error logging should be handled by Thanos itself, so this
patch turns off standard logging from the Azure library.

However, [due to the way logging is set up in the library][1], it will
always log errors to syslog, regardless of how `ShouldLog` is
configured. As such, and until that can be a configurable behavior, care
should be taken to handle that from the syslog side as well.

[1]: https://github.com/Azure/azure-storage-blob-go/blob/48358e1de5110852097ebbc11c53581d64d47300/azblob/zc_policy_request_log.go#L100-L102

Signed-off-by: Pedro Araujo <phcrva@gmail.com>