cmd: compact: clean partial / marked blocks concurrently (#3115)
* cmd: compact: clean partial / marked blocks concurrently

Clean partially uploaded and blocks marked for deletion concurrently
with the whole compaction/downsampling process. One iteration could
potentially take a few days so it should be nice to periodically clean
unneeded blocks in the background. Without this, there are huge spikes
in block storage usage. The spike's size depends on how long it takes to
complete one iteration.

The implementation of this is simple - factored out the deletion part
into a separate function. It is called at the end of an iteration +
concurrently if `--wait` has been specified. Add a mutex to protect from
concurrent runs. Delete blocks from the deletion mark map so that we
wouldn't try to delete same blocks twice or more.

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>

* *: update changelog, e2e tests

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>

* cmd: compact: fix according to comments

Remove "error" from the `error` and just directly call the function.

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>

* CHANGELOG: cleanups

Forgot to remove this part while solving conflicts.

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>

* CHANGELOG: update

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>

* CHANGELOG: clean whitespace

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>