Replaces errutil with tools/pkg/merrors; Fixes very ugly misuse of multi-error.
Previous multi-error implementation could cause very ugly bug of returnig empty multi-error
that should be treated as success not error by API, but if .Err() is not invoked it will be
used as non nil error.

There were 9 places where we had this bug in handler due to MultiError lib allowing to do so.

Signed-off-by: Bartlomiej Plotka <bwplotka@gmail.com>

# Conflicts:
#	pkg/block/fetcher.go
#	pkg/compact/compact.go
#	pkg/compact/compact_test.go
#	pkg/errutil/multierror.go
#	pkg/receive/handler.go
#	pkg/receive/handler_test.go