cmd/thanos/receive: fix dialopts TLS heuristic
Currently, the Thanos receive gRPC dialopts are configured to use TLS
when the remote-write HTTP server is configured with a TLS certificate.
This heuristic is not correct: just because the remote-write HTTP server
is exposed over TLS does not mean that the hashring gRPC server is using
TLS. This commit fixes the heuristic for enabling TLS in the gRPC client
for forwarding remote-write requests internally within a hashring.

Fixes: https://github.com/thanos-io/thanos/issues/4382.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>