compact: Add benchmark for #4084 (#4617)
* testing: Add a bucket that adds a delay for ops

Create a bucket that simulates (a non jittering) network RTT to be used
for system like benchmarks.

Signed-off-by: Holger Hans Peter Freyther <holger@moiji-mobile.com>

* compact: Add benchmark for GatherNoCompactionMarkFilter

Create a benchmark to show the benefit of adding concurrency
to the no compaction mark filter. Set-up an in-memory bucket
with a simulated (network) delay. In the future we can refactor
this a bit more to test more/all filters.

Fixes: #4084
Signed-off-by: Holger Hans Peter Freyther <holger@moiji-mobile.com>

* Update pkg/compact/compact_test.go

Co-authored-by: Giedrius Statkevičius <giedriuswork@gmail.com>
Signed-off-by: Holger Hans Peter Freyther <holger@moiji-mobile.com>

Co-authored-by: Giedrius Statkevičius <giedriuswork@gmail.com>