sidecar: not ready when Prometheus is unavailable
Set Sidecar to not be ready when Prometheus is unavailable. This avoids
problems when Prometheus is replaying WAL and Sidecar is "up" but then
Query hangs while waiting for a response from Prometheus/Sidecar. Also,
it gets shown as UP in the Stores page with no ext. labels during this
time that is confusing.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>