store: close bkt properly
Just a small fix: add bkt closing to the run group instead of closing it
at the of the function that only sets things up. In practice, this
didn't cause a bug because Close() is a noop. But as good Go
programmers, let's Close() the bucket properly at the end of Thanos
Store's lifetime.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>