cache: fix quickstart script, improve checking
Check that SelfURL also doesn't end with a slash since the galaxycache's
code checks whether the strings are equivalent when picking peers, and
peers cannot have slashes at the end so we need the same, exact format
of the URLs.

Fix quickstart script so that it wouldn't have slashes at the end.

Idea for the future: let's add smoke tests for `scripts/quickstart.sh`
to ensure smooth developer experience?

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>