makefile: Fix Makefile support for arm64
Currently the Makefile does not properly detect arm64 architecture
through uname -m. It assumes that all arm64 CPUs will be detected
as armv8.

This improves the detection by handling arm64 as a separate case.
It also formats the Makefile to use tabs instead of spaces in order
to prevent failures on OS X.

Signed-off-by: fpetkovski <filip.petkovsky@gmail.com>