Implement concurrent exemplar selects
Exemplars are currently fetched with a mixture of serial and concurrent
selects. Selects in multi-tsdb are done serially, while the exemplar
proxy fans out, waits for all responses and then forwards all exemplars
upstream.

This commit modifies the fanout mechanism to send exemplars concurrently
from both multi-tsdb and the exemplar proxy as soon as they are received.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>