Added Vertical Query Sharding to Query-Frontend (#5342)
* Update faillint to v1.10.0

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>

* Implement query sharding

This commit implements query sharding for grouping PromQL expressions.

Sharding is initiated by analyzing the PromQL and extracting
grouping labels. Extracted labels are propagated down to Stores which
partition the response by hashmoding all series on those labels.

If a query is shardable, the partitioning and merging process will be
initiated by the Query Frontend. The Query Frontend will make N distinct
queries across a set of Queriers and merge the results back before
presenting them to the user.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>

* First code review pass

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>

* Use sync pool to reuse sharding buffers

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>

* Add test for binary expression with constant

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>

* Include external labels in series sharding

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>