Ensure distinct replication (#5604)
* Ensure distinct replication

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>

* Precalculate hashring replicas

The ketama hashring algorithm can cause requests to be replicated
to the same node since one node can claim two subsequent sections
in the ring.

In addition to this, the replication algorithm in the receive handler
assumes that a single batch will always be replicated to one node.
While this is a good optimization for the hashmod algorithm,
it cannot be generalized for every other type of hashring.

This commit fixes both issues.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>
Signed-off-by: Prakul Jain <prakul.jain@udaan.com>