Route parse errors from sharding through normal api path
When the parser used in the querySharder encounters a parse error,
it returns that error back through to the API without the normal
api error wrapper.  This results in parse errors being sent back
with a 500 response code rather than a 400 response code, and the
UI (thanos ui, grafana) treats errors as a generic server error rather
than a client-side error.

This change breaks out of the sharding middleware and passes the request
along the chain to allow the same error handling to occur for parse errors.

Signed-off-by: Samuel Dufel <samuel.dufel@shopify.com>