Send sorted series from Receiver
Thanos Receivers currently send series in unsorted order, which
the store proxy from sorting on the fly using the k-way merge heap.
As a result, the querier has to do an explicit global sort before
deduplication.

This commit changes the series select call in the Receiver and Ruler
to request series in sorted order. It also adds a field to the info
message of each store which indicates whether the store sends series
in sorted order.

This field can be used for backwards compatibility once the global sort
is removed in the querier. We also have external stores which might
not send data in sorted order, for which this field will be set to
false.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>