Implement dedup-ready label sorting in stores
This commit adds functionality to all stores to sort series labels
in order appropriate for deduplication. This is done by propagating
replica labels from the Querier in the Series request. When all stores
send a hint that they support sorting labels, we can rely on the k-way
merge in the proxy and avoid an expensive global sort.

For backwards compatibility, each Store will send a hint in the
SeriesResponse to indicate that it supports this functionality.
The hint is sent at the beginning of the stream, and is the first
response that each store will send.

The Store Proxy will inspect the hints from each store and detect
cases where a store does not support resorting of labels. In that case
it will itself send a hint upstream that sorting is not supported.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>

# Conflicts:
#	pkg/store/bucket.go