Decide write quorum on the basis on number of succesfull replica requests
Instead of deciding on the basis of individual requests, this change takes into consideration the fact that there can be multiple requests per each replica number (for simplification I refer to these as 'replica groups'). For confirming quorum it is then decisive to assert whether ALL requests succeeded from AT LEAST the number of replica groups equal to the write quorom.

Signed-off-by: Matej Gera <matejgera@gmail.com>