Short-circuit chunk dedup in proxy
The store proxy deduplicates series by calculating a hash on the each
aggregate chunk. If at least one hash is not present in the dedup-map,
the entire series chunk is added to the response, including all aggregate
chunks.

Because of this, we can short-circuit deduping chunks and stop early if any
aggregate chunks is not previously seen.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>