Receive: reduce disruption when changing hashring
With the current implementation, whenever the hashrings config file changes
all receivers that detect any change will flush their TSDBs and stop ingestion
for a short period of time. This also leads to out of bounds samples since
a new head block will be created, and samples with the previous timestamp
will be rejected.

This commit reduces the need to flush TSDBs by comparing old hashrings to new
hashrings. If an endpoint was not part of a hashring that was modified, it will not
flush its TSDBs.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>