Fix quorum calculation for Ketama hashring
The quorum calculation is currently broken when using the Ketama
hashring. The reasons are explained in detail in issue
https://github.com/thanos-io/thanos/issues/5784.

This commit fixes quorum calculation by tracking successfull writes
for each individual time-series inside a remote-write request.

The commit also removes the replicate() method inside the Handler
and moves the entire logic of fanning out and calculating success
into the fanoutForward() method.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>