Fix mint and maxt for api/v1/series call
The storage querier currently uses select hints to set the start
and end time of the series select rpc. The same rpc is also used
for the api/v1/series API which does not produce any hints.

This commit changes the querier to use the mint and maxt from the
query itself when selecting series from stores.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>