Implement distributed query mode
This commit adds support for distributed query execution based on
the distributed mode of the Thanos PromQL engine.

Currently the mode is enabled through a hidden flag which will allow
us to run it in a test environment before releasing it as an experimental
feature. Another reason for having the flag hidden is that it allows
us to auto-detect the mode before the feature is released,
instead of having the user configure it explicitly.

The current version still lacks support for overriding query parameters,
such as resolution, store filtering etc. through the UI. Adding
support for those features will likely require an interface change
in the distributed engine.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>